+++ metrics-scala/src/main/scala/nl/grons/metrics4/scala/MetricBuilder.scala
@@ -26,7 +26,7 @@ import _root_.scala.concurrent.duration.FiniteDuration
 /**
  * Builds and registering metrics.
  */
-class MetricBuilder(val baseName: MetricName, val registry: MetricRegistry) {
+class MetricBuilder(val baseName: MetricName, val registry: MetricRegistry) extends DeprecatedMetricBuilder {
 
   private[this] val gauges: AtomicReference[Seq[DropwizardGauge[_]]] = new AtomicReference(Seq.empty)
 
@@ -34,10 +34,9 @@ class MetricBuilder(val baseName: MetricName, val registry: MetricRegistry) {
    * Registers a new gauge metric.
    *
    * @param name the name of the gauge
-   * @param scope the scope of the gauge or null for no scope
    */
-  def gauge[A](name: String, scope: String = null)(f: => A): Gauge[A] = {
-    wrapDwGauge(metricNameFor(name, scope), new DropwizardGauge[A] { def getValue: A = f })
+  def gauge[A](name: String)(f: => A): Gauge[A] = {
+    wrapDwGauge(metricNameFor(name), new DropwizardGauge[A] { def getValue: A = f })
   }
 
   /**
@@ -45,10 +44,9 @@ class MetricBuilder(val baseName: MetricName, val registry: MetricRegistry) {
    *
    * @param name the name of the gauge
    * @param timeout the timeout
-   * @param scope the scope of the gauge or null for no scope
    */
-  def cachedGauge[A](name: String, timeout: FiniteDuration, scope: String = null)(f: => A): Gauge[A] = {
-    wrapDwGauge(metricNameFor(name, scope), new DropwizardCachedGauge[A](timeout.length, timeout.unit) { def loadValue: A = f })
+  def cachedGauge[A](name: String, timeout: FiniteDuration)(f: => A): Gauge[A] = {
+    wrapDwGauge(metricNameFor(name), new DropwizardCachedGauge[A](timeout.length, timeout.unit) { def loadValue: A = f })
   }
 
   private def wrapDwGauge[A](name: String, dwGauge: DropwizardGauge[A]): Gauge[A] = {
@@ -61,37 +59,33 @@ class MetricBuilder(val baseName: MetricName, val registry: MetricRegistry) {
    * Creates a new counter metric.
    *
    * @param name the name of the counter
-   * @param scope the scope of the counter or null for no scope
    */
-  def counter(name: String, scope: String = null): Counter =
-    new Counter(registry.counter(metricNameFor(name, scope)))
+  def counter(name: String): Counter =
+    new Counter(registry.counter(metricNameFor(name)))
 
   /**
    * Creates a new histogram metric.
    *
    * @param name the name of the histogram
-   * @param scope the scope of the histogram or null for no scope
    */
-  def histogram(name: String, scope: String = null): Histogram =
-    new Histogram(registry.histogram(metricNameFor(name, scope)))
+  def histogram(name: String): Histogram =
+    new Histogram(registry.histogram(metricNameFor(name)))
 
   /**
    * Creates a new meter metric.
    *
    * @param name the name of the meter
-   * @param scope the scope of the meter or null for no scope
    */
-  def meter(name: String, scope: String = null): Meter =
-    new Meter(registry.meter(metricNameFor(name, scope)))
+  def meter(name: String): Meter =
+    new Meter(registry.meter(metricNameFor(name)))
 
   /**
    * Creates a new timer metric.
    *
    * @param name the name of the timer
-   * @param scope the scope of the timer or null for no scope
    */
-  def timer(name: String, scope: String = null): Timer =
-    new Timer(registry.timer(metricNameFor(name, scope)))
+  def timer(name: String): Timer =
+    new Timer(registry.timer(metricNameFor(name)))
 
   /**
    * Unregisters all gauges that were created through this builder.
@@ -104,6 +98,77 @@ class MetricBuilder(val baseName: MetricName, val registry: MetricRegistry) {
     })
   }
 
+  protected def metricNameFor(name: String): String = baseName.append(name).name
+}
+
+trait DeprecatedMetricBuilder { this: MetricBuilder =>
+
+  /**
+    * Registers a new gauge metric.
+    *
+    * @param name the name of the gauge
+    * @param scope (deprecated) the scope of the gauge or null for no scope
+    */
+  @deprecated("""Please use `gauge(name+"."+scope)(f)` instead. The scope parameter has been deprecated and will be removed in v5.0.0.""", "4.0.3")
+  def gauge[A](name: String, scope: String = null)(f: => A): Gauge[A] =
+    gauge(mergeScope(name, scope))(f)
+
+  /**
+    * Registers a new gauge metric that caches its value for a given duration.
+    *
+    * @param name the name of the gauge
+    * @param timeout the timeout
+    * @param scope (deprecated) the scope of the gauge or null for no scope
+    */
+  @deprecated("""Please use `cachedGauge(name+"."+scope)(f)` instead. The scope parameter has been deprecated and will be removed in v5.0.0.""", "4.0.3")
+  def cachedGauge[A](name: String, timeout: FiniteDuration, scope: String = null)(f: => A): Gauge[A] =
+    cachedGauge(mergeScope(name, scope), timeout)(f)
+
+  /**
+    * Creates a new counter metric.
+    *
+    * @param name the name of the counter
+    * @param scope (deprecated) the scope of the counter or null for no scope
+    */
+  @deprecated("""Please use `counter(name+"."+scope)(f)` instead. The scope parameter has been deprecated and will be removed in v5.0.0.""", "4.0.3")
+  def counter(name: String, scope: String = null): Counter =
+    counter(mergeScope(name, scope))
+
+  /**
+    * Creates a new histogram metric.
+    *
+    * @param name the name of the histogram
+    * @param scope (deprecated) the scope of the histogram or null for no scope
+    */
+  @deprecated("""Please use `histogram(name+"."+scope)(f)` instead. The scope parameter has been deprecated and will be removed in v5.0.0.""", "4.0.3")
+  def histogram(name: String, scope: String = null): Histogram =
+    histogram(mergeScope(name, scope))
+
+  /**
+    * Creates a new meter metric.
+    *
+    * @param name the name of the meter
+    * @param scope (deprecated) the scope of the meter or null for no scope
+    */
+  @deprecated("""Please use `meter(name+"."+scope)(f)` instead. The scope parameter has been deprecated and will be removed in v5.0.0.""", "4.0.3")
+  def meter(name: String, scope: String = null): Meter =
+    meter(mergeScope(name, scope))
+
+  /**
+    * Creates a new timer metric.
+    *
+    * @param name the name of the timer
+    * @param scope (deprecated) the scope of the timer or null for no scope
+    */
+  @deprecated("""Please use `timer(name+"."+scope)(f)` instead. The scope parameter has been deprecated and will be removed in v5.0.0.""", "4.0.3")
+  def timer(name: String, scope: String = null): Timer =
+    timer(mergeScope(name, scope))
+
+  private def mergeScope(name: String, scope: String): String =
+    if (scope == null) name else name + "." + scope
+
+  @deprecated("""Do not use metricNameFor, it is an internal API. This method will be removed in v5.0.0.""", "4.0.3")
   protected def metricNameFor(name: String, scope: String = null): String =
     baseName.append(name, scope).name
+
 }